moved: `handle_errors` func to log_config from drivers.py
